user_management/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
sqlalchemy==1.4.23
httpx==0.24.1
python-dotenv==0.19.0
email-validator==1.1.3 